Add a basic integration test skeleton.
Add a basic integration test skeleton to make it easier to extend the integration
with new features. Right now this just tests init and mocks out the client library,
but can be extended with actual interesting tests later.

Fix exception handling on startup so that it fails with retry instead of proceeding.