Add battery-level calculation for devices that report mV
This calculation is based on the L91 battery's datasheet that shows a distinct set of slopes to measure the approximate percentage of battery life left in devices that report 4000+ as the state.  This number is the mV of the 6 batteries in the battery powered Nest Protect and the 3 batteries as backup in the wired version.  I have tested this on 11 battery and 3 wired, and the results are pretty accurate based on the info I can see and the use I know each sensor gets.  Temperature sensors (for the thermostat) simply pass through as they did previously because it was already returning percentage for those.  In my tests, my batteries range from 100% to 27% with everything in between.  I also had one need replacement and it was sub 10% with the "battery low" state true as reported by the API.

Closes #67