Reduce timeout for subscription sessions to avoid lost connections (#425)
* Reduce timeout for subscription sessions to avoid lost connections

The timeout has been reduced from 3600 x 24 to 600, to operate as a heart beat. The total timeout was too long to handle well lost internet connections. With the reduced timeout, if the HA instance is disconnected from the internet and loose the callback connection, which is not caught by the aiohttp as an exception, the integration will re-initiate a connection as the timeout reached. 

The limit set to 600, as on previous versions where the integration was failing, it was reconnecting every 300 seconds. It is the double of that for not introducing excessive load on any server.

* Bump version to 0.4.1b1

---------

Co-authored-by: Mick Vleeshouwer <mick@imick.nl>